+++ src/rankify/InstanceBase.ts
@@ -11,29 +11,8 @@ import {
 } from "viem";
 import { ApiError, findContractDeploymentBlock, handleRPCError } from "../utils/index";
 import { getSharedSecret } from "@noble/secp256k1";
-
+import { gameStatusEnum } from "./types";
 import instanceAbi, { RankifyDiamondInstanceAbi } from "../abis/RankifyDiamondInstance";
-
-/**
- * Enum representing different states of a game instance
- * @public
- */
-export enum gameStatusEnum {
-  /** Game has been created but not opened for registration */
-  created = "Game created",
-  /** Game is open for player registration */
-  open = "Registration open",
-  /** Game is in progress */
-  started = "In progress",
-  /** Game is in its final turn */
-  lastTurn = "Playing last turn",
-  /** Game is in overtime */
-  overtime = "PLaying in overtime",
-  /** Game has finished */
-  finished = "Finished",
-  /** Game was not found */
-  notFound = "not found",
-}
 
 interface GameState extends ContractFunctionReturnType<typeof instanceAbi, "view", "getGameState"> {
   joinRequirements: ContractFunctionReturnType<typeof instanceAbi, "view", "getJoinRequirements">;
